fix: type conversion error for elapsed and moving time

+++ lib/domain/model/model_detailed_activity.dart
@@ -537,9 +537,9 @@ class SummaryGear {
 
 class SplitsMetric {
   double? distance;
-  int? elapsedTime;
+  double? elapsedTime;
   double? elevationDifference;
-  int? movingTime;
+  double? movingTime;
   int? split;
   double? averageSpeed;
   int? paceZone;
